requirements.txt
click>=8.1.3
requests>=2.28.2
python-dotenv>=1.0.0
rich>=13.3.5
pydantic>=2.4.0
typer>=0.9.0
websockets>=12.0
prompt_toolkit>=3.0.20
mcp[cli]>=0.1.0
pytest>=7.0.0  # Para testes
fastapi>=0.103.0
uvicorn>=0.23.0
openai>=1.0.0
pyyaml>=6.0.1
# Dependências opcionais para suporte a CrewAI
crewai>=0.11.2
mcp-run>=0.3.0